[REFACTOR] create stack
